Delete: Api Manager, Client 통합
